Move taxons to the constants directory
